requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
openai
python-dotenv
